requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
numpy==1.21.0
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
gunicorn==20.1.0
django-cors-headers==3.7.0